LayoutTests/http/tests/security/redirect-BLOCKED-to-localURL.html

<html>
<body>
<iframe src="http://127.0.0.1:8000/security/resources/redir.php?url=file:///tmp/LayoutTests/http/tests/security/resources/file-redirect-target.html"></iframe></br>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
</script>
<p>This attempts to open a redirect link to a file URL, which should be blocked.</p>
</body>
</html>
