safari-600.7-branch/LayoutTests/http/tests/security/redirect-BLOCKED-to-localURL-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: file-redirect-target.html

This attempts to open a redirect link to a file URL, which should be blocked.
